Added Shapes and Falling
The boids now have a direction they are pointing and move in that direction while moving. There is some messiness with that but will hopefully be ironed out later in a fix that is coming to another section.

On click now the boids will get shot and fall down to the bottom of the screen. The dead pile will be added later on.